refactor: split mission-control-page into task-card, task-list, and thread-panel modules
